requirements.txt
amqp==2.1.4

appdirs==1.4.3

billiard==3.5.0.2

celery==4.0.2

Django==1.11
djangorestframework == 3.7.7
django-filter
django-cors-headers==2.1.0

django-haystack==2.6.0

jieba==0.38

kombu==4.0.2

packaging==16.8
#pkg-resources==0.0.0

psycopg2==2.7.3.2

pymongo==3.4.0

PyMySQL==0.7.10

pyparsing==2.2.0

pytz==2016.10

#redis==2.10.5

requests==2.18.4

six==1.10.0

vine==1.1.3

Whoosh==2.7.4

category==0.9.10

simplejson == 3.13.2
